apps/api/image-to-video/requirements.txt
accelerate
diffusers[torch]
fastapi
huggingface_hub
moviepy
pillow
python-multipart
transformers
uvicorn[standard]
